Apply configuration once, to reduce bootup latency
Problem
=======

Using "systemd-analyze" to inspect the bootup latency, we can quickly
see the "delphix-platform" adds considerable latency to bootup time.

Since the "delphix-mgmt" service depends on "delphix-platform", any time
required to execute the "delphix-platform" service directly affects the
latency for the "delphix-mgmt" to come up after a reboot.

Here's an example from a system without this change:

    $ systemd-analyze blame | head -n2
        1min 20.155s delphix-mgmt.service
             18.660s delphix-platform.service

From this, we see it takes over 18s to execute the "delphix-platform"
service; this translates to 18 seconds of additional downtime for the
"delphix-mgmt" service.

Solution
========

To address this issue, this change modifies the "delphix-platform"
service to only run the ansible configuration a single time. This will
allow the service to configure the system during the first boot (which
will obviously entail some bootup latency), but will prevent the service
from running the same ansible configuration on subsequent boots
(eliminating the additional bootup latency caused by this service for
these subsequent boots).

Here's an example from a system with this change:

    $ sudo systemd-analyze blame | head -n2
        1min 22.222s delphix-mgmt.service
             10.496s rtslib-fb-targetctl.service

    $ sudo systemd-analyze blame | grep delphix-platform
               123ms delphix-platform.service

In this example, the "delphix-platform" now adds a very minimal amount
of additionally latency to the bootup time, because the service doesn't
have to re-apply the ansible configuration that's already been applied.

Caveats
=======

The caveat to this approach being, "delphix-platform" service will no
longer automatically apply new configuration changes after an upgrade of
the Delphix appliance.

Since we actually do want to re-apply the ansible configuration after an
upgrade (e.g. because the ansible logic may have been changed due to the
upgrade), the logic/scripts we use to orchestrate the upgrade will need
to be aware of this solution, and force it to run (e.g. by using
"systemctl reload" to remove the /etc/delphix-platform/ansible-done file
as part of the upgrade).