Create and preserve "/var/log/delphix-upgrade" directory
When performing an upgrade via the virtualization API, the application
will attempt to create a file in "/var/log/delphix-upgrade" prior to
executing any of the upgrade scripts. This file creation currently fails
because the "/var/log/delphix-upgrade" directory will not exist at that
point (that directory is created by the upgrade scripts).

This change fixes the problem by creating this directory via the
delphix-platform service (which runs on first boot, prior to the
virtualization application running), and also updates the associated
cron entry to ensure the directory is not removed.