Fix "override.conf" for SSH service
The change in commit 1cf09aad4a4769c6cfaea7dda79232dc20bebc06 was not
updated after we landed commit a221e9579a206336a86045f9fc97784b36677d8d.
We need to move the "override.conf" file for the SSH service to adhere
to the new directory structure; which this change is doing.