Add new script to read version.info properties
This change adds a new script that can be used to more easily extract
the value of properties contained in an upgrade image's "version.info"
file. In addition to reading the property from the "version.info" file,
this script is responsible for also verifying the signature and checksum
for that file and upgrade image.

This enables consumers to more easily retrieve these properties, without
having to commit to fully unpacking the image.

Here's a few examples:

    $ get-property-from-image internal-dev.upgrade.tar.gz VERSION
    2019.8.27.4

    $ get-property-from-image -s internal-dev.upgrade.tar.gz VERSION
    2019.8.27.4

    $ get-property-from-image internal-dev.upgrade.tar.gz UNKNOWN
    get-property-from-image: failed to get property 'UNKNOWN'

    $ get-property-from-image internal-dev.upgrade.tar.gz MINIMUM_REBOOT_OPTIONAL_VERSION
    5.3.0.0

    $ get-property-from-image internal-dev.upgrade.tar.gz
    get-property-from-image: too few arguments specified
    Usage: get-property-from-image [-s] <image> <property>

    $ get-property-from-image VERSION internal-dev.upgrade.tar.gz
    get-property-from-image: The upgrade image must be a '.upgrade.tar.gz' file

    $ get-property-from-image
    get-property-from-image: too few arguments specified
    Usage: get-property-from-image [-s] <image> <property>