Various improvements to allow migration stress testing in the mgmt stack
We add a new "migration complete" flag file that is set by the mgmt stack
when the migration upgrade is completed. This simplifies the logic in
migrationctl to determine when all migration steps are completed.

migrationctl resume functionality is improved to take into account services
that are auto-restarted on failure. Note that this doesn't apply to the mgmt
stack as auto-restart functionality is now disabled in the mgmt stack for
the duration of migration; this makes stress testing easier and also
improves the debugging experience by avoiding restarting multiple times on
the same failures.