Sporadic failures when loading the crash kernel in low memory due to KASLR
On our current kernel version (v4.15) the reservation for the
crash kernel is done within the first 896MB of memory by default
in one contiguous chunk. This is problematic for us because we
currently need 512MB for our crash kernel which not only is pretty
big for such a small window but also the allocation is guaranteed
to fail whenever the running kernel (not the crash kernel) is
randomly placed within that window by KASLR.

This patch splits the reservation for the crash kernel into two
parts with one of them being in high memory (e.g. outside the
required region of the first 896MBs). This provides more
flexibility on how we can size it and making the allocations for
the reservation more likely to succeed.

For more background info on the actual bug refer to DLPX-65761.
For more background info on the design of this refer to the
block comment of this patch.