DLPX-66005 Unquiesced VDBs cause domain0 to fail to mount due to a bug in zfs inherit sharenfs
zfs inherit sharenfs automatically mounts each dataset which has
sharenfs set. However the parent datasets will not be mounted if
they do not have sharenfs set, which will result in the creation
of mount directories in the wrong file systems and will later
cause zfs mount -a to fail with "directory not empty". As a
work around to that bug, we make sure to call zfs mount -a first.