Automatically determine image variant if not specified
This change optimizes the UX of "download-latest-image" script, such
that the user does not necessarily have to specify the image variant to
be downloaded.

This is useful for the case where the script is run on an existing
Delphix appliance, and the user wants the script to download the image
variant that matches the appliance it's being run on.

In this case, the script can (and now does, with this change) detect the
variant of the appliance, and download the corresponding image, without
forcing the user to specify which variant to download.