Add new "download-release-image" script (#341)
This adds another script to ease the burden of downloading upgrade
images. We already have the "download-latest-image" script that can be
used to download not yet released upgrade images, and this change
compliments that script by adding a "download-released-image" script
that can be used to download already released upgrade images.