Update and rename 20-override.conf to aws-override.conf
The 20-override.conf file has been renamed to a more meaningful name "aws-override.conf". The new name communicates the intent that this file is specific to AWS behavior. 
An empty "ExecStartPre=" directive resets any previous ExecStartPre= directives. This behavior is not needed and was added due to some misunderstanding on my part. Hence, I have removed this empty "ExecStartPre=" directive.