Update update_netplan_to_mac_address
Modify the location of the interface to mac address map file from /var/tmp/delphix-upgrade to /etc. This is being done because we found out in testing that "/var/tmp/delphix-mgmt/interface_to_macaddress_map.out" is not available (likely does not exist) when the systemd-networkd daemon is started in the initramfs phase. The "/etc/interface_to_macaddress_map.out" path is available. It seems this is a result of only specific paths getting built into the initramfs image.